+++ packages/neon-auth-next/src/index.ts
@@ -1,8 +1,8 @@
 export { toNextJsHandler } from "./handler";
 export { neonAuthMiddleware } from "./middleware";
-import { createAuthClient as createNeonAuthClient, BetterAuthReactAdapter } from "@neondatabase/neon-auth"
+import { createAuthClient as createNeonAuthClient, BetterAuthReactAdapter } from "@neondatabase/auth"
 
-export * from "@neondatabase/neon-auth";
+export * from "@neondatabase/auth";
 
 export const createAuthClient = () => {
   // @ts-expect-error - for nextjs proxy we do not need the baseUrl