+++ packages/neon-auth-next/CHANGELOG.md
@@ -9,20 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bump @neondatabase/neon-auth to 0.1.0-alpha.6
+- Bump @neondatabase/auth to 0.1.0-alpha.6
 
 
 ## [0.1.0-alpha.11] - 2025-12-03
 
 ### Changed
 
-- Bump @neondatabase/neon-auth to 0.1.0-alpha.5
+- Bump @neondatabase/auth to 0.1.0-alpha.5
 
 ## [0.1.0-alpha.4] - 2025-12-03
 
 ### Changed
 
-- Bump @neondatabase/neon-auth to 0.1.0-alpha.4
+- Bump @neondatabase/auth to 0.1.0-alpha.4
 
 ## [Unreleased]